refactor(src): Object -> Module로 구조 변경
함수를 담고 있는 Object에서 Module형으로 구조 변경
이를 통해 함수 이름을 명확히 선언하고 사용할 수 있음
Parser, Validator, Printer -> parseInput, validateData, printData
