--- src/Printer.js
@@ -1,15 +1,0 @@
-import { Console } from '@woowacourse/mission-utils';
-
-const Printer = {
-  currentStatus(carList) {
-    carList.forEach(({ name, position }) => {
-      Console.print(`${name} : ${'-'.repeat(position)}`);
-    });
-  },
-
-  blankLine() {
-    Console.print('');
-  },
-};
-
-export default Printer;
