+++ src/utils/runCarRacing.js
@@ -1,11 +1,11 @@
-import Printer from '../Printer.js';
+import { printBlankLine, printCurrentStatus } from '../printData.js';
 
 // (Car[], number) => void
 const runCarRacing = (carList, turn) => {
   for (let i = 0; i < turn; i++) {
     actionAllCarOneTurn(carList);
-    Printer.currentStatus(carList);
-    Printer.blankLine();
+    printCurrentStatus(carList);
+    printBlankLine();
   }
 };
